+++ crates/ruma-client-api/src/r0/room/create_room.rs
@@ -11,7 +11,7 @@ use ruma_events::{
     },
     AnyInitialStateEvent,
 };
-use ruma_identifiers::{RoomId, RoomVersionId, UserId};
+use ruma_identifiers::{RoomId, RoomName, RoomVersionId, UserId};
 use ruma_serde::{Raw, StringEnum};
 use serde::{Deserialize, Serialize};
 
@@ -58,7 +58,7 @@ ruma_api! {
         /// If this is included, an `m.room.name` event will be sent into the room to indicate
         /// the name of the room.
         #[serde(skip_serializing_if = "Option::is_none")]
-        pub name: Option<&'a str>,
+        pub name: Option<&'a RoomName>,
 
         /// Power level content to override in the default power level event.
         #[serde(skip_serializing_if = "Option::is_none")]