+++ crates/ruma-events/tests/initial_state.rs
@@ -1,7 +1,8 @@
 use std::convert::TryFrom;
 
 use matches::assert_matches;
-use ruma_events::{room::name::RoomName, AnyInitialStateEvent, InitialStateEvent};
+use ruma_events::{AnyInitialStateEvent, InitialStateEvent};
+use ruma_identifiers::RoomNameBox;
 use serde_json::json;
 
 #[test]
@@ -13,7 +14,7 @@ fn deserialize_initial_state_event() {
         }))
         .unwrap(),
         AnyInitialStateEvent::RoomName(InitialStateEvent { content, state_key})
-        if content.name == Some(RoomName::try_from("foo".to_owned()).unwrap())
+        if content.name == Some(RoomNameBox::try_from("foo").unwrap())
             && state_key.is_empty()
     );
 }