+++ crates/ruma-events/tests/stripped.rs
@@ -2,10 +2,10 @@ use std::convert::TryFrom;
 
 use js_int::uint;
 use ruma_events::{
-    room::{join_rules::JoinRule, name::RoomName, topic::TopicEventContent},
+    room::{join_rules::JoinRule, topic::TopicEventContent},
     AnyStateEventContent, AnyStrippedStateEvent, StrippedStateEvent,
 };
-use ruma_identifiers::{mxc_uri, user_id};
+use ruma_identifiers::{mxc_uri, user_id, RoomNameBox};
 use serde_json::{from_value as from_json_value, json, to_value as to_json_value};
 
 #[test]
@@ -96,7 +96,7 @@ fn deserialize_stripped_state_events() {
     let event = from_json_value::<AnyStrippedStateEvent>(name_event).unwrap();
     match event {
         AnyStrippedStateEvent::RoomName(event) => {
-            assert_eq!(event.content.name, Some(RoomName::try_from("Ruma".to_owned()).unwrap()));
+            assert_eq!(event.content.name, Some(RoomNameBox::try_from("Ruma").unwrap()));
             assert_eq!(event.state_key, "");
             assert_eq!(event.sender.to_string(), "@example:localhost");
         }